pyproject.toml
[project]
name = "qd-hr-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.15.0",
    "fastapi>=0.115.14",
    "fastembed-gpu>=0.7.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-cohere>=0.4.4",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langserve>=0.3.1",
    "pandas>=2.3.1",
    "pdfminer-six>=20250506",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.3",
    "ragas>=0.2.15",
    "redis>=6.2.0",
    "sse-starlette>=2.3.6",
    "unstructured[docx]>=0.18.2",
    "uvicorn>=0.35.0",
]
